Rename `pathfilter` to `fsfilter`
